LayoutTests/imported/w3c/web-platform-tests/streams/readable-byte-streams/general.dedicatedworker-expected.txt


PASS getReader({mode: "byob"}) throws on non-bytes streams 
PASS ReadableStream with byte source can be constructed with no errors 
PASS ReadableStream with byte source: Construct and expect start and pull being called 
PASS ReadableStream with byte source: No automatic pull call if start doesn't finish 
PASS ReadableStream with byte source: Construct with highWaterMark of 0 
PASS ReadableStream with byte source: getReader(), then releaseLock() 
FAIL ReadableStream with byte source: getReader() with mode set to byob, then releaseLock() ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: Test that closing a stream does not release a reader automatically 
FAIL ReadableStream with byte source: Test that closing a stream does not release a BYOB reader automatically ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: Test that erroring a stream does not release a reader automatically 
FAIL ReadableStream with byte source: Test that erroring a stream does not release a BYOB reader automatically ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: releaseLock() on ReadableStreamReader with pending read() must throw 
PASS ReadableStream with byte source: Automatic pull() after start() 
PASS ReadableStream with byte source: Automatic pull() after start() and read() 
FAIL ReadableStream with byte source: autoAllocateChunkSize assert_equals: pull() must have been invoked once expected 1 but got 0
FAIL ReadableStream with byte source: Mix of auto allocate and BYOB promise_test: Unhandled rejection with value: object "TypeError: ReadableStreamBYOBRequest respond() is not implemented"
PASS ReadableStream with byte source: Automatic pull() after start() and read(view) 
PASS ReadableStream with byte source: enqueue(), getReader(), then read() 
PASS ReadableStream with byte source: Push source that doesn't understand pull signal 
PASS ReadableStream with byte source: read(), but pull() function is not callable 
FAIL ReadableStream with byte source: read(view), but pull() function is not callable ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: enqueue() with Uint16Array, getReader(), then read() 
FAIL ReadableStream with byte source: enqueue(), read(view) partially, then read() ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: getReader(), enqueue(), close(), then read() 
PASS ReadableStream with byte source: enqueue(), close(), getReader(), then read() 
PASS ReadableStream with byte source: Respond to pull() by enqueue() 
PASS ReadableStream with byte source: Respond to pull() by enqueue() asynchronously 
FAIL ReadableStream with byte source: read(view), then respond() ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view), then respond() with a transferred ArrayBuffer ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view), then respond() with too big value ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: respond(3) to read(view) with 2 element Uint16Array enqueues the 1 byte remainder ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: enqueue(), getReader(), then read(view) ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: enqueue(), getReader(), then cancel() (mode = not BYOB) 
FAIL ReadableStream with byte source: enqueue(), getReader(), then cancel() (mode = BYOB) ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: getReader(), read(view), then cancel() ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: cancel() with partially filled pending pull() request promise_test: Unhandled rejection with value: object "TypeError: ReadableStreamBYOBReader is not implemented"
FAIL ReadableStream with byte source: enqueue(), getReader(), then read(view) where view.buffer is not fully covered by view ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: Multiple enqueue(), getReader(), then read(view) ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: enqueue(), getReader(), then read(view) with a bigger view ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: enqueue(), getReader(), then read(view) with a smaller views ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: enqueue() 1 byte, getReader(), then read(view) with Uint16Array ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: enqueue() 3 byte, getReader(), then read(view) with 2-element Uint16Array promise_test: Unhandled rejection with value: object "TypeError: ReadableStreamBYOBReader is not implemented"
FAIL ReadableStream with byte source: read(view) with Uint16Array on close()-d stream with 1 byte enqueue()-d must fail ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: A stream must be errored if close()-d before fulfilling read(view) with Uint16Array ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: Throw if close()-ed more than once 
PASS ReadableStream with byte source: Throw on enqueue() after close() 
FAIL ReadableStream with byte source: read(view), then respond() and close() in pull() ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view) with Uint32Array, then fill it by multiple respond() calls ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: read() twice, then enqueue() twice 
FAIL ReadableStream with byte source: Multiple read(view), close() and respond() ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: Multiple read(view), big enqueue() ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: Multiple read(view) and multiple enqueue() ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view) with passing undefined as view must fail ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view) with zero-length view must fail ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view) with passing an empty object as view must fail ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: Even read(view) with passing ArrayBufferView like object as view must fail ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: read() on an errored stream 
PASS ReadableStream with byte source: read(), then error() 
FAIL ReadableStream with byte source: read(view) on an errored stream ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: read(view), then error() ReadableStreamBYOBReader is not implemented
PASS ReadableStream with byte source: Throwing in pull function must error the stream 
PASS ReadableStream with byte source: Throwing in pull in response to read() must be ignored if the stream is errored in it 
FAIL ReadableStream with byte source: Throwing in pull in response to read(view) function must error the stream ReadableStreamBYOBReader is not implemented
FAIL ReadableStream with byte source: Throwing in pull in response to read(view) must be ignored if the stream is errored in it ReadableStreamBYOBReader is not implemented
FAIL ReadableStreamBYOBReader can be constructed directly ReadableStreamBYOBReader is not implemented
FAIL ReadableStreamBYOBReader constructor requires a ReadableStream argument ReadableStreamBYOBReader is not implemented
FAIL ReadableStreamBYOBReader constructor requires an unlocked ReadableStream ReadableStreamBYOBReader is not implemented
FAIL ReadableStreamBYOBReader constructor requires a ReadableStream with type "bytes" ReadableStreamBYOBReader is not implemented
